+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 * Add accordion to the template sets on the deploy page
 * Add additional status reporting for tasks in the Deploy Log
 * Add filter to Templates Page
-* Add delete all button to Templates Page
+* Add disabled templatesets to table on Templates Page
+* Add install button to templatesets
 
 ## Fixed
 * Stop printing empty strings to the console